requirements.txt
asgiref==3.8.1
astroid==3.3.9
colorama==0.4.6
dill==0.4.0
django==5.2
django-filter==25.1
djangorestframework==3.16.0
isort==6.0.1
mccabe==0.7.0
numpy==2.2.5
platformdirs==4.3.7
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pylint==3.3.6
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing-extensions==4.13.2
tzdata==2025.2
